Add files via upload
program files for the vending machine part of the assignment.
Snack.java defines what snacks are and should have (name, price, quantity)

StateOfVendingMachine.java is an interface for the different states of the vending machine (idle, payment, dispense)

IdleState.java defines the idle state, where the user can select a snack, which then moves to the payment state

PaymentState.java defines the payment state, where the user inserts money to pay. If the cost is met, then moves to dispense state.

DispenseState.java defines the payment state, where the machine dispenses the payed-for snack item

SnackDispenseHandler.java defines the handler for chain of responsibility for the machine. When vending snacks, the machine goes through the chain until the right snack handler is reached, then the snack is dispensed.

VendingMachine.java defines the actual vending machine. Calls and changes states depending on the action being performed, and calls uses the dispense state to call the handler when dispensing snacks.

VendingMachineDriver.java is the main driver program

TestVendingMachine.java contains the unit tests